# [THRUST] Faster multi dimensional argsort by segmented sort