--- a/python/tvm/topi/cuda/sort.py
+++ b/python/tvm/topi/cuda/sort.py
@@ -409,68 +409,6 @@ def sort_by_key_ir(
     )
 
 
-def argsort_nms_thrust(data, valid_count, axis=-1, is_ascend=1, dtype="float32"):
-    """Performs sorting along the given axis and returns an array of indicies
-    having same shape as an input array that index data in sorted order.
-
-    Parameters
-    ----------
-    data: tvm.te.Tensor
-        The input array.
-
-    valid_count : tvm.te.Tensor, optional
-        The number of valid elements to be sorted.
-
-    axis : int, optional
-        Axis long which to sort the input tensor.
-
-    is_ascend : boolean, optional
-        Whether to sort in ascending or descending order.
-
-    dtype : string, optional
-        DType of the output indices.
-
-    Returns
-    -------
-    out : tvm.te.Tensor
-        The output of this function.
-    """
-    ndim = len(data.shape)
-    if axis < 0:
-        axis = ndim + axis
-    if axis != ndim - 1:
-        # Prepare for sorting along axis -1.
-        axes = swap(list(range(ndim)), axis)
-        data = transpose(data, axes)
-
-    data_buf = tvm.tir.decl_buffer(data.shape, data.dtype, "data_buf", data_alignment=8)
-    valid_count_buf = tvm.tir.decl_buffer(
-        valid_count.shape, valid_count.dtype, "valid_count_buf", data_alignment=4
-    )
-    out_bufs = [
-        tvm.tir.decl_buffer(data.shape, data.dtype, "value_buf", data_alignment=8),
-        tvm.tir.decl_buffer(data.shape, "int32", "indices_buf", data_alignment=8),
-    ]
-    out = te.extern(
-        [data.shape, data.shape],
-        [data, valid_count],
-        lambda ins, outs: tvm.tir.call_packed(
-            "tvm.contrib.thrust.sort_nms", ins[0], ins[1], outs[0], outs[1], is_ascend
-        ),
-        in_buffers=[data_buf, valid_count_buf],
-        out_buffers=out_bufs,
-        dtype=[data.dtype, "int32"],
-        name="nms_argsort_gpu",
-        tag="nms_argsort_gpu",
-    )
-
-    if axis != ndim - 1:
-        axes = swap(list(range(ndim)), axis)
-        out = [transpose(o, axes) for o in out]
-
-    return out[1]
-
-
 def sort(data, axis=-1, is_ascend=1):
     """Performs sorting along the given axis and returns an array of
     sorted values with the same shape as the input data.
@@ -602,7 +540,7 @@ def argsort(data, axis=-1, is_ascend=1, dtype="float32"):
     return out
 
 
-def argsort_thrust(data, valid_count=None, axis=-1, is_ascend=1, dtype="float32"):
+def argsort_thrust(data, axis=-1, is_ascend=1, dtype="float32"):
     """Performs sorting along the given axis and returns an array of indicies
     having same shape as an input array that index data in sorted order.
 
@@ -610,9 +548,6 @@ def argsort_thrust(data, valid_count=None, axis=-1, is_ascend=1, dtype="float32"
     ----------
     data: tvm.te.Tensor
         The input array.
-
-    valid_count : tvm.te.Tensor, optional
-        The number of valid elements to be sorted.
 
     axis : int, optional
         Axis long which to sort the input tensor.
@@ -628,11 +563,7 @@ def argsort_thrust(data, valid_count=None, axis=-1, is_ascend=1, dtype="float32"
     out : tvm.te.Tensor
         The output of this function.
     """
-    if valid_count is not None:
-        out = argsort_nms_thrust(data, valid_count, axis, is_ascend, dtype)
-    else:
-        out = topk_thrust(data, 0, axis, "indices", is_ascend, dtype)
-    return out
+    return topk_thrust(data, 0, axis, "indices", is_ascend, dtype)
 
 
 def schedule_sort(outs):