Current implementation of thrust argsort, when given multi dimensional inputs to sort along the inner most axis, is very inefficient: it does `n_iter` calls to thrust sort. See

https://github.com/apache/tvm/blob/bad149ed8a555444d813537608ee5cea9e95e97e/src/runtime/contrib/thrust/thrust.cu#L50-L65

When the outer dimension is large, the performance of thrust argsort is far from optimal. In particular, the thrust numbers shown in the TIR mergesort PR https://github.com/apache/tvm/pull/7099 do not reflect the true performance thrust can achieve. 

This PR replaces `n_iter` calls to thrust argsort with one **segmented sort by key**. Since thrust doesn't provide API to do segmented sort, I used a neat back-to-back stable-sort-by-key trick explained in https://groups.google.com/forum/#!topic/thrust-users/BoLsxO6b4FY. My implementation is a bit more complicated because we need to do segmented sort **by key**, not just segmented sort. 

Here are the numbers I get using the same benchmark script used in https://github.com/apache/tvm/pull/7099, measured on GTX 1070 ti. When the outer dimension is small (like 2, 2, 2000 case), my change makes it slower due to the overhead from two calls to `stable_sort_by_key`. But other than that, it is much faster than one we have now.

Also, I removed `tvm.contrib.thrust.sort_nms` and `argsort_nms_thrust`, since they are not used anymore.

please review @kazum @Laurawly 
(cc @mbrookhart when you are back, this should be exciting for you!) 

| Shape         | current thrust | after this PR | TIR mergesort  |
|---------------|---------|--------|-------|
| (2000, 2, 2)  | 0.17    | 0.29   | 1.63  |
| (2, 2000, 2)  | 0.17    | 0.31   | 1.62  |
| (2, 2, 2000)  | 0.16    | 0.30   | 1.62  |
| (4000, 2, 2)  | 0.18   | 0.30  | 3.80  |
| (2, 4000, 2)  | 0.18   | 0.30   | 3.24  |
| (2, 2, 4000)  | 0.17    | 0.30   | 3.92  |
| (2, 12000, 2) | 0.52  | 0.59  | 12.96 |
| (2, 2, 12000) | 0.57   | 0.58   | 11.40 |
| (12000, 2, 2) | 0.63  | 0.59   | 11.77 |
| (2000, 8, 8)  | 2.55  | 0.90   | 4.08 |
| (8, 2000, 8)  | 2.43  | 0.81   | 3.56  |
| (8, 8, 2000)  | 2.47   | 0.88   | 3.02  |
| (4000, 8, 8)  | 2.55  | 1.43   | 9.37  |
| (8, 4000, 8)  | 2.54  | 1.42   | 9.68 |
| (8, 8, 4000)  | 2.53  | 1.41   | 6.14  |
| (12000, 8, 8) | 14.72 | 3.32  | 39.37 |
| (8, 12000, 8) | 13.79 | 3.24  | 40.27 |
| (8, 8, 12000) | 13.29 | 3.19  | 25.87 |