@trevor-m 2021-01-04
Nice! Have you also looked at CUB's [DeviceSegmentedRadixSort::SortPairsDescending](https://nvlabs.github.io/cub/structcub_1_1_device_segmented_radix_sort.html#abbba6639f928bebd19435f185ea10618) ? It sounds like it is exactly what you need with no tricks required. It's used by some fast NMS implementations such as [TensorRT](https://github.com/NVIDIA/TensorRT/tree/master/plugin/batchedNMSPlugin).